Created hockey and dev flavor so no hockeyapp things pop up during development
