+++ sample/src/dev/java/net/gini/switchsdk/sample/BaseActivity.java
@@ -1,0 +1,7 @@
+package net.gini.switchsdk.sample;
+
+
+import android.support.v7.app.AppCompatActivity;
+
+public class BaseActivity extends AppCompatActivity {
+}
